[mypyc] Refactor: extract Mapper from mypyc.genops
